Completed rebase of profiler- onto fault-tolerance
This commit concludes the rebase of the performance profiling code
(first part of master thesis) onto the fault-tolerance work (second
part of master thesis).

All code that does actual measurements has been removed! Only the
framework in Profiler.{hpp,cpp} and the test cases remain.

It is planned to use the measurement code to profile the performance
of the fault-tolerant mechanismns.